requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
keras==2.13.1
tensorflow-cpu==2.13.0
Pillow==9.5.0
numpy==1.24.3
joblib==1.3.2
scikit-learn==1.2.2
pandas==2.0.3
Werkzeug==2.0.2
